+++ docs/changelog.rst
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+* :release:`v0.4.1 <2015-08-11>`
+* :feature:`0` Add tests for simoa.stats.online_variance
+
 * :release:`v0.4.0 <2015-08-10>`
 * :feature:`0` New documentation and internal refactoring